requirements.txt
streamlit==1.40.0
pandas==2.2.2
pillow==11.0.0
pdf2image==1.17.0
cryptography==43.0.3
openai==1.52.1
xlsxwriter==3.1.1
